Initial Version
Packaged APB VIP in a separate folder

Packaged APB VIP in a separate folder

Packaged APB VIP in a separate folder

Fixed clock generation

The clock was getting generated for a fixed number of cycles. Changed
that to an infinite loop.

Build the driver and the sequencer only if the agent is active

Initial Version

It is now a separate verification IP

Added apb as a separate VIP as a submodule

Compile the testbench without the debug flag

APB Interface is not parameterized

Fixed the SHA monitor to figure out the SHA3 rate from WRITE phase

Do not reuse access_enum from APB in SHA3

This is required to keep SHA3 and APB functionalities independent of
each other

Create parameterized sequence item directly from the UVM factory

Do not use setup and hold delays in testbench and design

Use bool variable is_write in apb_seq_item instead of access_enum

Since we are using new eUVM, no need to compile trace.d separately